Tried making the discovery into an AsyncTask
