+++ Hestia/app/src/main/java/hestia/UI/dialogs/AddDeviceDialog.java
@@ -3,6 +3,8 @@ package hestia.UI.dialogs;
 import android.app.AlertDialog;
 import android.app.Dialog;
 import android.content.DialogInterface;
+import android.net.nsd.NsdManager;
+import android.net.nsd.NsdServiceInfo;
 import android.os.AsyncTask;
 import android.os.Bundle;
 import android.support.v4.app.FragmentManager;
@@ -13,18 +15,23 @@ import android.view.WindowManager;
 import android.widget.ArrayAdapter;
 import android.widget.AutoCompleteTextView;
 import android.widget.Toast;
+
 import com.rugged.application.hestia.R;
+
 import java.io.IOException;
+import java.net.InetAddress;
 import java.util.ArrayList;
+
 import hestia.backend.ServerCollectionsInteractor;
 import hestia.backend.exceptions.ComFaultException;
 import hestia.backend.models.RequiredInfo;
 
 /**
-* This class opens the dialog to enter the collection name and plugin name.
-* It then sends this to the networkHandler which tries to get the required info.
-* If this works it consecutively opens a new dialog for the other info.
-* @see EnterRequiredInfoDialog
+ * This class opens the dialog to enter the collection name and plugin name.
+ * It then sends this to the networkHandler which tries to get the required info.
+ * If this works it consecutively opens a new dialog for the other info.
+ *
+ * @see EnterRequiredInfoDialog
  */
 
 public class AddDeviceDialog extends HestiaDialog {
@@ -63,7 +70,7 @@ public class AddDeviceDialog extends HestiaDialog {
 
                 // Negative Button
                 .setNegativeButton("Cancel", new DialogInterface.OnClickListener() {
-                    public void onClick(DialogInterface dialog,	int which) {
+                    public void onClick(DialogInterface dialog, int which) {
                         // Do something else
                     }
 
@@ -108,14 +115,14 @@ public class AddDeviceDialog extends HestiaDialog {
         final String collection = collectionField.getText().toString();
         final String pluginName = pluginField.getText().toString();
 
-         new AsyncTask<Object, String, RequiredInfo>() {
+        new AsyncTask<Object, String, RequiredInfo>() {
             @Override
             protected RequiredInfo doInBackground(Object... params) {
                 RequiredInfo info = null;
                 try {
                     info = serverCollectionsInteractor.getRequiredInfo(collection, pluginName);
                 } catch (IOException e) {
-                    Log.e(TAG,e.toString());
+                    Log.e(TAG, e.toString());
                     String exceptionMessage = "Could not connect to the server";
                     publishProgress(exceptionMessage);
                 } catch (ComFaultException comFaultException) {
@@ -128,17 +135,17 @@ public class AddDeviceDialog extends HestiaDialog {
                 return info;
             }
 
-             @Override
-             protected void onProgressUpdate(String... exceptionMessage) {
-                 Toast.makeText(getContext(), exceptionMessage[0], Toast.LENGTH_SHORT).show();
-             }
+            @Override
+            protected void onProgressUpdate(String... exceptionMessage) {
+                Toast.makeText(getContext(), exceptionMessage[0], Toast.LENGTH_SHORT).show();
+            }
 
             @Override
             protected void onPostExecute(RequiredInfo info) {
-                if(info != null){
+                if (info != null) {
                     EnterRequiredInfoDialog fragment = EnterRequiredInfoDialog.newInstance();
                     fragment.setData(info, serverCollectionsInteractor);
-                    if(fragmentManager == null){
+                    if (fragmentManager == null) {
                     }
                     fragment.show(fragmentManager, "dialog");
                 }
@@ -154,7 +161,7 @@ public class AddDeviceDialog extends HestiaDialog {
                 try {
                     collections = serverCollectionsInteractor.getCollections();
                 } catch (IOException e) {
-                    Log.e(TAG,e.toString());
+                    Log.e(TAG, e.toString());
                     String exceptionMessage = "Could not connect to the server";
                     publishProgress(exceptionMessage);
                 } catch (ComFaultException comFaultException) {
@@ -190,7 +197,7 @@ public class AddDeviceDialog extends HestiaDialog {
                 try {
                     plugins = serverCollectionsInteractor.getPlugins(collection);
                 } catch (IOException e) {
-                    Log.e(TAG,e.toString());
+                    Log.e(TAG, e.toString());
                     String exceptionMessage = "Could not connect to the server";
                     publishProgress(exceptionMessage);
                 } catch (ComFaultException comFaultException) {
@@ -211,7 +218,7 @@ public class AddDeviceDialog extends HestiaDialog {
             @Override
             protected void onPostExecute(ArrayList<String> plugins) {
                 adapterPlugins.clear();
-                if(plugins != null) {
+                if (plugins != null) {
                     adapterPlugins.addAll(plugins);
                 }
             }
